expose SIP and RTP port to all interface
why: we must bind to 0.0.0.0, otherwise it's really difficult to use
softphone on the same host. Anyways this config make it possible to pass
the REGISTER step.
Also, if we doesn't specify udp AND tcp, but both are available on the
container, docker will take tcp by default. That's why we need to expose
both explicitly

Next step is to add wazo-agid to successfully dialing *10